Add Quote component and refactor structure of the files
