+++ src/components/shared/Quote.jsx
