+++ src/index.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Navbar from './components/navbar/Navbar';
+// import Navbar from './components/navbar/Navbar';
+import Home from './components/home/Home';
 
 ReactDOM.render(
   <React.StrictMode>
-    <Navbar />
+    <Home />
   </React.StrictMode>,
   document.getElementById('root'),
 );
